+++ fbEngin/TestChara.cpp
@@ -23,7 +23,7 @@ void CTestChara::Start()
 
 void CTestChara::Update()
 {
-	m_Animation.Update(SINSTANCE(CTimer)->DeltaTime());
+	m_Animation.Update((float)SINSTANCE(CTimer)->DeltaTime());
 }
 
 void CTestChara::Render()
